extracted some functions from index.js
